WIP: Improve all public APIs by dropping (using Context)  (#538)
* removed most of (using Context) from public-facing APIs
* split rpc from core
* Try to speed up builds by @prolativ
* increase IT logging to debug deadlock on CI
* drop most sonatype:snapshots and remove all autogenerated project.scala in IT from git
---------

+++ core/src/main/scala/besom/internal/ProtobufUtil.scala
@@ -22,7 +22,7 @@ object ProtobufUtil:
     extension (i: Int) def asValue: Value = Value(Kind.NumberValue(i))
 
   given ToValue[Long] with
-    extension (l: Long) def asValue: Value = Value(Kind.NumberValue(l))
+    extension (l: Long) def asValue: Value = Value(Kind.NumberValue(l.toDouble))
 
   given ToValue[String] with
     extension (s: String) def asValue: Value = Value(Kind.StringValue(s))
